fixato un problema per cui i posti di una proiezione venivano salvati come float, impedendo l'avvio del programma

+++ src/main/java/Proiezioni.java
@@ -129,7 +129,7 @@ public class Proiezioni {
      */
     public float getPrezzo(){ return this.prezzo;}
 
-    public float getPosti(){return this.posti;}
+    public int getPosti(){return this.posti;}
 
     // setter
